Order error when price is 1 uah (#1359)
* changed links in pugs

* fixed negative price with certficate

* fixed tests

* fixed tests

+++ modules/order/order.graphql.js
@@ -20,8 +20,8 @@ type Order {
   fixedExchangeRate: Float
   promoCodeId: String
   certificateId: String
-  itemsPriceWithDiscount: [Int]
-  itemsDiscount: [Int]
+  itemsPriceWithDiscount: [Float]
+  itemsDiscount: [Float]
 }
 
 enum PaymentStatusEnum {
